+++ src/main/java/com/backend/timeforpizza/timeforpizzabackend/repository/CuriosityRepository.java
@@ -2,9 +2,8 @@ package com.backend.timeforpizza.timeforpizzabackend.repository;
 
 import com.backend.timeforpizza.timeforpizzabackend.model.Curiosity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("curiosityRepository")
-public interface CuriosityRepository extends JpaRepository<Curiosity, Integer> {
+public interface CuriosityRepository extends JpaRepository<Curiosity, Long> {
 }
