+++ src/main/java/com/backend/timeforpizza/timeforpizzabackend/repository/IngredientRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository("ingredientRepository")
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
     List<Ingredient> findByRecipe(Recipe recipe);
     void deleteByRecipe(Recipe recipe);
 }
