+++ src/main/java/com/backend/timeforpizza/timeforpizzabackend/service/IngredientService.java
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -53,7 +52,7 @@ public class IngredientService {
         ingredientRepository.delete(ingredient);
     }
 
-    public void deleteIngredientById(Integer ingredientId) {
+    public void deleteIngredientById(Long ingredientId) {
         ingredientRepository.deleteById(ingredientId);
     }
